Reorganize Minimap reference
Bug, zoom and size but not working again
The goal is to make minimap asset an independent asset that can be
exported alone.
Put all assets in the minimap folder.
minimap was referencing avatar through POI_reference hub. To reduce
interdependency between tools, the avatar reference changed by directly
accessing the attached object, as minimap manager is on the avatar