Package Cleanup
Created a FullPackage folder that contains all of the necessary Assets.
Inside that folder there is a BaseScene where we can add features and
functionality.  There is also a Prefab of the _DontDestroyOnLoad object
that can be dragged into a scene to get the avatar and all of the added
widgets.