Added in new characters/animations
Added some basic character models that we used at the ARL for avatars.
Also imported animations that we can look at to improve the avatars.

The avatars can be used for the "Drop Avatar" option, with the
possibility for having a drop down menu to select the avatar you want to
drop.