Character Drop and Menu Toggle
I added in the button to show/hide the menu in the bottom left corner.
It now shows the settings button and the drop character button on the
left side of the screen.  I started implementing the ability to drop
avatars, but will need to solve the issue of multiple cameras before I
can finish implementation.