Added in InputMap scripts
Added a folder in the project called InputMap and placed all the script
needed for setting up the InputManager in there.  We will need to decide
how we want to handle looking for and loading input files, and whether
we want the InputManager to do it on its own or have another script
handle loading.