Projector for selecting local wander radius
When dropping a new character in local wander mode, the first right
click selects the local wander center, then as the user moves the mouse
the radius will be shown as a black circle on the ground.  The second
right click sets the wander radius and drops the character.  Still need
to implement the radius selector in the CharacterOptionsPanel.