Created a generic way to open one menu
Created a new class WidgetMenu that implements a function ToggleMenu().
The WidgetCanvasManager now handles opening and closing menus to ensure
that only one menu is open at a time.  When creating a new menu, there
should be a script on the button for the new menu that inherits from
WidgetMenu.  It should then implement the ToggleMenu function to handle
opening and closing the menu.  The button should then call
OpenMenu(WidgetMenu) from a reference to the widget canvas manager. with
a reference to the WidgetMenu script on the button.