Reset references for POI buttons
The POI buttons were working for me, but I reset them incase there was
an issue after moving the scripts around.  Also, I fixed an error in the
TP_Camera when hitting the toggleCamera key.  I wasn't away of this
until an error popped up while i was naming a POI.  Will need to look
into disabling this.