Cleaned up package
Moved all necessary items under the Full Package folder.  The
FullPackage folder can be exported as its own package and imported into
another project.  To keep the size of this package reasonable, I left
the character prefabs in a separate folder.  The characters will need to
be brought into the new project separately.