Created a scene with menu transitions
Added a scene in the UITransitions folder that has a panel showing a
sliding in/out, fading in/out, and growing/shrinking animation.  They
are set up to play the animation and end up back in the center of the
screen.  Just click one of the buttons at the bottom of the scene to see
the animation.