Setup Gaze Input for our menus
Tested with minimap buttons.  I moved the display settings file so that
it wouldnt get loaded.  That way there is only one camera, and I set it
up so that this camera can trigger the GazeInput events.  The buttons
are clicked after you have looked at them for 2 seconds.  The circle
crosshair shows where you are pointing.