Bug fixes on UI and character dropper (#6)
* updated pins and characters

cleaned up scene

* Fixed character drop point

Characters are now correctly dropping at the saved location

* Char info radius 2 digits

Limited the character info panel to only display the wander radius to 2
digits after the decimal

* Fixed new characters going to bookmark

added in a line to initialize new characters to no poiDestination

* Testing and Code Check for Bookmark

small changes to scripts.  Tested characters going into and out of
Bookmark wander mode.  Does not display in charinfopanel.  Checked code
to make sure it is working as intended.

* Fixed characters walking outside Patrol circle

Patrol characters now select their new destination from within the
circle instead of a square.

* Fixed idle time and animator

Characters in idle mode should no longer inch around in a circle.  Also
fixed the idleTime, which is the amount of time the characters will idle
between reaching their destination and picking a new one.

* Recentered the Widget3DObject

recentered the widget3dobject so that characters are dropped in the
correct place.  this change did not seem to get saved in the previous
commit.

* Added in POIEdit window Open/Close functions

Added functions to contain the code for Opening and Closing the POIEdit
window.

Also added in code for filling in the POIInfo fields so that there
wouldn't be issues with fields being out of order or incorrectly set.

Fixed positioning of the POIEditWindow.

* Fixed bugs and references with CustomBookmark

Fixed issues with the CustomBookmark button and switched to reference
the new Open and Close functions for the POIEdit window.

* Fixed issue with WidgetRoot/Config transitions.

the transitions, sliding off screen, will only be triggered if there in
not already a transition running on that object.  This prevents the menu
from transitioning too far or starting a new transition in the middle of
an existing one.

* Fixed loaded characters and poidestination

fixed error where the poidestination on characters was getting set
before the user gave the character one

* Fixed character movement after charinfo open

Characters should no longer move after character info is open.  Opening
character info sets a flag that prevents movement until a new mode is
set on the character.

* Added IConUtilities

Added an IConUtilities class that contains a SetPopUpWindow function
that takes in a recttransform and return the closest position to the
current mouse position that keeps the entire recttransform on the
screen.

* Added UI position check for POI marker canvas