Added in manual width for menus and fixed pause
POI edit mode no longer changes the timescale.  The timescale code was
in two places in the script.

Added in a manual width field to the Flexible Draggable script so that
we can hardcode the width of certain menus.  This is currently being
used to correctly size the POIMenu to include all of its submenus.