Added UI_Extensions package
Was able to find a UI_Extensions package that contains a lot of useful
UI improvements and scripts.

Use the Bound ToolTip scripts to set up tooltips that appear for the
name of the poi's when hovering over the buttons.

Need to only show the tooltip if there is overflow and add it to more UI
objects.