UI improvements and bug fixes(#9)
* Added in Drag and Drop POIMenu

Added in a drag and drop script on the poi menu.

Created another function in UIUtilities that places a recttransform on a
grid on the screen.  this will be expanded to accommodate other
anchor/pivot configs and handle multiple menu objects not overlapping.

* ControlUtilities update

Added in ControlUtilities class with Pause and UnPause functions.

* Menu State Manager cleanup

Cleaned up the POIMenuStateManager script.

Added/removed/changed functions to improve the workflow.

* POIEdit window transition cleanup

consolidated the code for opening/closing the POIEdit window into
functions in the POIMenuStateManager script.

Commented code in POIMenuStateManager.

Removed some unnecessary scripts.

* Fixed MarkerInfoCanvas Delete

Fixed the Delete button on the MarkerInfoCanvas so that it properly
deletes the markers now.

The hierarchy of the pin changed so the reference to the POIInfo script
changed.

* Set up UI Collisions template

Started integrating the 2D physics collision detection for preventing ui
overlap.

Editted the FlexibleDraggable script to incorporate the physics
components, Rigidbody2d and BoxCollider2D.

Added FlexibleDraggable to more menus in POI system.

Started testing settings to get desired behaviour.

* Corner calculations for UI elements

Started integrating code to use the corner calculations to tell when a
menu item is on screen.

Will also use the colliders attached to the menu objects to check
against a stored list of ui colliders to make sure no two menus are
overlapping.

still in development

* Use corners to keep menus on screen

Use the corners of the menu object to make sure it stays on screen while
its being dragged.

* Removed timescale change from poi menu

* Added in manual width for menus and fixed pause

POI edit mode no longer changes the timescale.  The timescale code was
in two places in the script.

Added in a manual width field to the Flexible Draggable script so that
we can hardcode the width of certain menus.  This is currently being
used to correctly size the POIMenu to include all of its submenus.

* Fixed issue with POI Exit button not working

* Added in overlap prevention

Use the OnCollisionEnter2D and OnCollisionExit2D functions on the
BoxCollider2D to prevent moving one menu over another one.

* Added Flexible Draggable to more menus

Added the Flexible Draggable script to the settings menus and the
character drop menus.

Fixed an issue with the display of settings files.

* Cleaned up code and commented

Added comments to different scripts.

Prevented camera rotation while dragging a menu.

* Added UI_Extensions package

Was able to find a UI_Extensions package that contains a lot of useful
UI improvements and scripts.

Use the Bound ToolTip scripts to set up tooltips that appear for the
name of the poi's when hovering over the buttons.

Need to only show the tooltip if there is overflow and add it to more UI
objects.

* Updated tooltip to only show when there is overflow

The tooltip will now display for POI's whose name does not fit in the
button.

* Added tooltip to widget settings menu

Tooltip now works for the descriptions in the widget settings menu

* Settings UI upgrade

* Added some comments to BoundTooltipItem

* Implementing new SettingsManager

Added in new SettingsManager script with functions and classes for
saving/loading/setting various widget, avatar, and keybinding settings.

* Comments and Menu integration

Added functions for filling in the various menus and updating the
settings from them.

Added comments some existing scripts.

Modified the menus to include fields for variables that the user can
change.  Also set input fields to allow the correct entry types.

* Finished adding in Menu transitions for Settings

The settings menus now properly transition between states and allow
altering of settings.

* Updated tooltip to include info, added poi save

* Tooltip and Edit mode changes

Added some functionality to the tooltip for proper display.

Added in a global EditModeManager to handle disabled the other widgets.

* Character Info and EditMode updates

Edit mode should now transition without error.

Character Info should now work as an edit mode.

Character Info should only open if the mouse is pressed down and
released on a character.

* UI tweak

* Testing for draggable window fix

* Improvements to DraggableObject

Improved the collisions on dragging menu objects around.  Still not
perfect through.

* Submenus for POI and Sunlight widgets open on sides

Submenus for POI and Sunlight will now open on the right or left to keep
them on the screen.

* Updated FlexibleDraggable with new scheme

Does not worry about keeping the menu on the screen.  If the dragged
menu is released over another menu, it is returned to its position at
the start of the drag.

* Added tooltip timers for display and hiding

Tooltip will now wait 2 seconds before displaying.

Tooltip will disappear 5 seconds after displaying.

* Fixed tooltip bug

Tooltip should now display and disappear correctly.

* Adjusted widgettransition time and distance

made sure the widgettransitions were getting all of the menus off the
screen.

* Bug Fixes

Fixed issue with CharInfoPanel display.

Settings button now properly turns red when selected.

Sunlight widget transitions fixed.

* Pins open when mouse down and up on same pin

POI pins and character should now only open their respective menus if
the mousedown and mouseup events occur over the same character/pin.

* Fixed issue with characters turning

Fixed a bug where characters would play their turning animation when
they should be idle.

* UI tweak sunlight widget

1. fixed close button not toggling to correct icon
2. fixed close button shifting due to horizontal layout group
3. optimized button click logic
4. new tooltip sprite

* same as last commit

* Simple TP_Camera fixes

* Prevent Keybindings being null

Keybindings will not update from the menu if the InputField is null.

* Fixed raycast lock issues and character colliders

Changed sizes of the character colliders to more accurately represent
the models.

Updated the RaycastLock so that it is locked when a menu enters edit
mode.